+++ packages/moleculer-db-adapter-mongoose/examples/models/posts.js
@@ -18,6 +18,10 @@ let PostSchema = new Schema({
 	},
 	author: {
 		type: Schema.ObjectId
+	},
+	status: {
+		type: Boolean,
+		default: true
 	}
 
 }, {