+++ packages/moleculer-db/src/memory-adapter.js
@@ -86,7 +86,7 @@ class MemoryDbAdapter {
 	 */
 	find(filters) {
 		return new Promise((resolve, reject) => {
-			this.doFiltering(filters).exec((err, docs) => {
+			this.createCursor(filters).exec((err, docs) => {
 				/* istanbul ignore next */
 				if (err)
 					return reject(err);
@@ -142,7 +142,7 @@ class MemoryDbAdapter {
 	 */
 	count(filters = {}) {
 		return new Promise((resolve, reject) => {
-			this.doFiltering(filters).exec((err, docs) => {
+			this.createCursor(filters).exec((err, docs) => {
 				/* istanbul ignore next */
 				if (err)
 					return reject(err);
@@ -184,7 +184,7 @@ class MemoryDbAdapter {
 	 * @memberof MemoryDbAdapter
 	 */
 	updateMany(query, update) {
-		return this.db.update(query, update, { multi: true, returnUpdatedDocs: true }).then(res => res[1]);
+		return this.db.update(query, update, { multi: true }).then(res => res[0]);
 	}
 
 	/**
@@ -256,7 +256,7 @@ class MemoryDbAdapter {
 	 * @param {Object} params 
 	 * @returns {Query}
 	 */
-	doFiltering(params) {
+	createCursor(params) {
 		if (params) {
 			let q;